Rename binaryAnalysis to memoryAccessCounter
This also does some code cleanup and creates an explicit build target
for the example.
