+++ CMakeLists.txt
@@ -63,7 +63,7 @@ add_executable(InterestingProgram common/mutatees/InterestingProgram.cpp)
 add_subdirectory(codeCoverage)
 add_subdirectory(unstrip)
 add_subdirectory(instrumentAFunction)
-add_subdirectory(binaryAnalysis)
+add_subdirectory(memoryAccessCounter)
 add_subdirectory(instrumentMemoryAccess)
 add_subdirectory(retee)
 add_subdirectory(maxArg)
