+++ memoryAccessCounter/CMakeLists.txt
@@ -1,0 +1,8 @@
+project(memoryAccessCounter LANGUAGES CXX)
+
+add_executable(counter counter.cpp)
+target_include_directories(counter PRIVATE ${DYNINST_INCLUDE_DIR})
+target_link_libraries(counter dyninstAPI instructionAPI)
+
+# A dummy target to allow building this example by itself
+add_custom_target(memoryAccessCounter DEPENDS counter)
